requirements.txt
asgiref==3.7.2
Babel==2.12.1
beautifulsoup4==4.12.2
bootstrap4==0.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
distlib==0.3.6
Django==4.2.1
django-bootstrap4==23.1
django-braces==1.15.0
django-cleanup==8.0.0
django-crispy-forms==2.0
docopt==0.6.2
filelock==3.12.0
virtualenvwrapper-win==1.2.7
wheel==0.38.4
yarg==0.1.9
